requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
numpy>=1.21.0
Pillow>=9.0.0
albumentations>=1.3.0
tensorboard>=2.12.0
matplotlib>=3.5.0
pandas>=1.5.0
scikit-learn>=1.0.0
tqdm>=4.65.0
GPUtil>=1.4.0  # For GPU monitoring
psutil>=5.9.0  # For system monitoring
requests>=2.31.0  # For API requests
python-dotenv>=1.0.0  # For environment variables
beautifulsoup4>=4.12.0
huggingface-hub>=0.19.0
datasets>=2.14.0
h5py>=3.8.0 